requirements.txt
absl-py==2.3.1
altair==5.5.0
astunparse==1.6.3
attrs==25.4.0
backtesting==0.6.5
beautifulsoup4==4.14.2
blinker==1.9.0
bokeh==3.8.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.3
curl_cffi==0.13.0
flatbuffers==25.9.23
frozendict==2.4.6
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.12.0
libclang==18.1.1
llvmlite==0.44.0
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.3
multitasking==0.0.12
namex==0.1.0
narwhals==2.10.0
newsapi-python==0.2.7
nltk==3.9.2
numba==0.61.2
numpy==2.2.6
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
pandas-ta==0.4.71b0
peewee==3.18.2
pillow==12.0.0
platformdirs==4.5.0
plotly==6.3.1
protobuf==6.33.0
pyarrow==22.0.0
pycparser==2.23
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
soupsieve==2.8
streamlit==1.51.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.0
xyzservices==2025.10.0
yfinance==0.2.66
